ai-engine/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pandas
scikit-learn
pydantic
